requirements.txt
Flask==2.3.3
openai==1.3.0
Pillow==10.0.1
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
urllib3==1.26.16
Werkzeug==2.3.7
Jinja2==3.1.2